@@ -16,26 +16,31 @@
     "test": "jest"
   },
   "devDependencies": {
-    "@microsoft/tsdoc": "^0.12.16",
-    "@types/jest": "^24.0.24",
-    "@types/minimatch": "^3.0.3",
-    "@types/node": "^13.1.0",
-    "@types/webpack": "^4.41.0",
-    "jest": "^24.9.0",
-    "ts-jest": "^24.2.0",
-    "tslint": "^5.20.1",
-    "typescript": "^3.7.4",
-    "webpack": "^4.41.4",
-    "webpack-recompilation-simulator": "^3.0.0"
+    "@microsoft/tsdoc": "^0.13.2",
+    "@types/find-cache-dir": "^3.2.1",
+    "@types/jest": "^27.0.2",
+    "@types/minimatch": "^3.0.5",
+    "@types/node": "^16.10.2",
+    "@types/webpack": "^5.28.0",
+    "@types/webpack-sources": "^3.2.0",
+    "jest": "^27.2.4",
+    "ts-jest": "^27.0.5",
+    "tslint": "^6.1.3",
+    "typescript": "^4.5.0-dev.20211001",
+    "webpack": "^5.55.1",
+    "webpack-recompilation-simulator": "^3.2.0",
+    "webpack-sources": "^3.2.1",
+    "tslib": "^2.3.1"
   },
   "dependencies": {
-    "free-tex-packer-core": "^0.2.9",
+    "find-cache-dir": "^3.3.2",
+    "free-tex-packer-core": "^0.3.4",
     "minimatch": "^3.0.4",
-    "readdir-enhanced": "^5.1.1",
-    "webpack-log": "^3.0.1"
+    "readdir-enhanced": "^6.0.4",
+    "spark-md5": "^3.0.1"
   },
   "peerDependencies": {
-    "webpack": "^4.3.7"
+    "webpack": "^5.55.1"
   },
   "files": [
     "dist",
